Provide main() function if configured to do so
We add config options for enabling a main() function provided by our Redis glue
code. This supports running both the server and the client. Given that a VM may
run as a Redis server but can also be a client to another Redis server, a
Redis-on-Unikraft instance can run with both server and client enabled.
Moreover, we can even connect the client to the local server.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>