Bump version from 7.0.5 to 7.0.11 and change patches
Because we use musl, several patches that changed pthreads calls
to work with newlib are no longer needed and have been removed.
Removed a patch that would remove an `#include` directive from
the code as that header is now necessary. Removed a patch that
would add a relative import path to a header, as that path is
part of `CINCLUDES`. Added a patch that removes a header which
causes compilation errors.

Signed-off-by: Ioan-Teodor Teugea <ioan_teodor.teugea@stud.acs.upb.ro>